# docs/openapi.yaml
openapi: 3.1.0
info:
  title: ReviewHelm API
  version: 1.0.0
  description: |
    REST API contract for ReviewHelm backend.
    Base path is /api/v1.
servers:
  - url: https://api.reviewhelm.com/api/v1
    description: Production
  - url: http://localhost:3000/api/v1
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /me:
    get:
      tags: [User]
      summary: Get current user
      responses:
        '200':
          description: Current user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/preferences:
    get:
      tags: [User]
      summary: Get user preferences
      responses:
        '200':
          description: User preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [User]
      summary: Update user preferences
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PreferencesPatchRequest'
      responses:
        '200':
          description: Updated preferences
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Preferences'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /me/api-keys/anthropic:
    get:
      tags: [Keys]
      summary: Get Anthropic key status for current user
      responses:
        '200':
          description: Anthropic key status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnthropicKeyStatus'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    put:
      tags: [Keys]
      summary: Upsert encrypted Anthropic key for current user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AnthropicKeyUpsertRequest'
      responses:
        '204':
          description: Key saved
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    delete:
      tags: [Keys]
      summary: Delete Anthropic key for current user
      responses:
        '204':
          description: Key deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
  /sessions/{sessionId}:
    get:
      tags: [Sessions]
      summary: Get session
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    patch:
      tags: [Sessions]
      summary: Update session metadata
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateSessionRequest'
      responses:
        '200':
          description: Updated session
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Session'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Sessions]
      summary: Delete session
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '204':
          description: Session deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /sessions/{sessionId}/summary:
    get:
      tags: [Sessions]
      summary: Get computed session summary
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionSummary'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /comment-profiles:
    get:
      tags: [Comment Profiles]
      summary: List comment drafting style profiles
      responses:
        '200':
          description: Profile list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentProfileListResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    post:
      tags: [Comment Profiles]
      summary: Create a new comment drafting profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateCommentProfileRequest'
      responses:
        '201':
          description: Created profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /comment-profiles/{profileId}:
    patch:
      tags: [Comment Profiles]
      summary: Update a comment profile
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateCommentProfileRequest'
      responses:
        '200':
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CommentProfile'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
    delete:
      tags: [Comment Profiles]
      summary: Delete a comment profile
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Profile deleted
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /comment-profiles/{profileId}/activate:
    post:
      tags: [Comment Profiles]
      summary: Set active comment profile
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Active profile updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  activeProfileId:
                    type: string
                required: [activeProfileId]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
...
  /usage/summary:
    get:
      tags: [Usage]
      summary: Get usage summary for month
      parameters:
        - in: query
          name: month
          schema:
            type: string
            pattern: '^\\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-03'
      responses:
        '200':
          description: Usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UsageSummary'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /usage/by-feature:
    get:
      tags: [Usage]
      summary: Get usage breakdown by feature for month
      parameters:
        - in: query
          name: month
          schema:
            type: string
            pattern: '^\\d{4}-(0[1-9]|1[0-2])$'
            example: '2026-03'
      responses:
        '200':
          description: Feature usage breakdown
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/FeatureCostBreakdown'
                required: [items]
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /usage/sessions/{sessionId}:
    get:
      tags: [Usage]
      summary: Get usage summary for one session
      parameters:
        - $ref: '#/components/parameters/SessionIdParam'
      responses:
        '200':
          description: Session usage summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionUsageSummary'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '404':
          $ref: '#/components/responses/NotFoundError'
  /usage/budget:
    get:
      tags: [Usage]
      summary: Get budget and guard configuration
      responses:
        '200':
          description: Budget config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetConfig'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
    patch:
      tags: [Usage]
      summary: Update budget and guard configuration
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BudgetConfigPatchRequest'
      responses:
        '200':
          description: Updated budget config
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BudgetConfig'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
  /usage/reset:
    post:
      tags: [Usage]
      summary: Reset usage counters for current user
      responses:
        '204':
          description: Usage reset
        '401':
          $ref: '#/components/responses/UnauthorizedError'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    SessionIdParam:
      in: path
      name: sessionId
      required: true
      schema:
        type: string
        format: uuid
...
  responses:
    UnauthorizedError:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    NotFoundError:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ValidationError:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            requestId:
              type: string
              format: uuid
          required: [code, message]
      required: [error]
    StackId:
      type: string
      enum:
        - java-protobuf
        - js-ts-react-node
        - go
        - terraform-hcl
        - swift-objc
        - web-devops-config
        - python
        - ruby
        - lua
        - c-lang
    ChecklistMode:
      type: string
      enum: [review, polish]
    Severity:
      type: string
      enum: [blocker, major, minor, nit]
    Verdict:
      type: string
      enum: [looks-good, needs-attention, na, skipped]
    ConfidenceLevel:
      type: integer
      enum: [1, 2, 3, 4, 5]
    ClaudeModel:
      type: string
      enum: [haiku, sonnet, opus]
...
    AiFeature:
      type: string
      enum: [learn, deep-dive, comment-drafter]
    UserProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        displayName:
          type: string
        createdAt:
          type: string
          format: date-time
      required: [id, email, createdAt]
    Preferences:
      type: object
      properties:
        aiModel:
          $ref: '#/components/schemas/ClaudeModel'
        defaultSeverityFilter:
          type: array
          items:
            $ref: '#/components/schemas/Severity'
        antiBiasMode:
          type: boolean
        fontSize:
          type: string
          enum: [small, medium, large]
        codeBlockTheme:
          type: string
          enum: [dark, light]
        autoExportPdf:
          type: boolean
        activeCommentStyleProfileId:
          type: string
          nullable: true
      required:
        - aiModel
        - defaultSeverityFilter
        - antiBiasMode
        - fontSize
        - codeBlockTheme
        - autoExportPdf
    PreferencesPatchRequest:
      type: object
      properties:
        aiModel:
          $ref: '#/components/schemas/ClaudeModel'
        defaultSeverityFilter:
          type: array
          items:
            $ref: '#/components/schemas/Severity'
        antiBiasMode:
          type: boolean
        fontSize:
          type: string
          enum: [small, medium, large]
        codeBlockTheme:
          type: string
          enum: [dark, light]
        autoExportPdf:
          type: boolean
        activeCommentStyleProfileId:
          type: string
          nullable: true
      additionalProperties: false
    AnthropicKeyStatus:
      type: object
      properties:
        configured:
          type: boolean
        tokenHint:
          type: string
          nullable: true
        updatedAt:
          type: string
          format: date-time
          nullable: true
        keySource:
          type: string
          enum: [byok, platform]
          nullable: true
        keyVersion:
          type: integer
          nullable: true
      required: [configured]
    AnthropicKeyUpsertRequest:
      type: object
      properties:
        apiKey:
          type: string
          minLength: 10
      required: [apiKey]
      additionalProperties: false
...
    UpdateSessionRequest:
      type: object
      properties:
        title:
          type: string
      required: [title]
      additionalProperties: false
    ItemResponse:
      type: object
      properties:
        verdict:
          $ref: '#/components/schemas/Verdict'
        confidence:
          $ref: '#/components/schemas/ConfidenceLevel'
        notes:
          type: string
        draftedComment:
          type: string
      required: [verdict, confidence]
...
    Session:
      type: object
      properties:
        id:
          type: string
          format: uuid
        mode:
          $ref: '#/components/schemas/ChecklistMode'
        stackId:
          allOf:
            - $ref: '#/components/schemas/StackId'
          nullable: true
        title:
          type: string
        itemResponses:
          type: object
          additionalProperties:
            $ref: '#/components/schemas/ItemResponse'
        sessionNotes:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        isComplete:
          type: boolean
      required:
        - id
        - mode
        - title
        - itemResponses
        - sessionNotes
        - createdAt
        - updatedAt
        - isComplete
...
    SessionScores:
      type: object
      properties:
        coverage:
          type: integer
        confidence:
          type: integer
        issuesBySeverity:
          type: object
          properties:
            blocker:
              type: integer
            major:
              type: integer
            minor:
              type: integer
            nit:
              type: integer
          required: [blocker, major, minor, nit]
        totalIssues:
          type: integer
        itemsResponded:
          type: integer
        applicableItems:
          type: integer
      required:
        - coverage
        - confidence
        - issuesBySeverity
        - totalIssues
        - itemsResponded
        - applicableItems
    FeatureCostBreakdown:
      type: object
      properties:
        feature:
          $ref: '#/components/schemas/AiFeature'
        calls:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        costUsd:
          type: number
      required: [feature, calls, inputTokens, outputTokens, costUsd]
    SessionUsageSummary:
      type: object
      properties:
        calls:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        costUsd:
          type: number
        byFeature:
          type: array
          items:
            $ref: '#/components/schemas/FeatureCostBreakdown'
      required: [calls, inputTokens, outputTokens, costUsd, byFeature]
    SessionSummary:
      type: object
      properties:
        scores:
          $ref: '#/components/schemas/SessionScores'
        lowConfidenceItems:
          type: array
          items:
            type: object
            additionalProperties: true
        sessionUsage:
          $ref: '#/components/schemas/SessionUsageSummary'
      required: [scores, lowConfidenceItems, sessionUsage]
...
    CommentProfile:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tone:
          type: string
        strictness:
          type: integer
        verbosity:
          type: integer
        includePraise:
          type: boolean
        includeActionItems:
          type: boolean
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tone
        - strictness
        - verbosity
        - includePraise
        - includeActionItems
        - createdAt
        - updatedAt
    CreateCommentProfileRequest:
      type: object
      properties:
        name:
          type: string
        tone:
          type: string
        strictness:
          type: integer
          minimum: 1
          maximum: 5
        verbosity:
          type: integer
          minimum: 1
          maximum: 5
        includePraise:
          type: boolean
        includeActionItems:
          type: boolean
      required: [name, tone, strictness, verbosity]
      additionalProperties: false
    UpdateCommentProfileRequest:
      type: object
      properties:
        name:
          type: string
        tone:
          type: string
        strictness:
          type: integer
          minimum: 1
          maximum: 5
        verbosity:
          type: integer
          minimum: 1
          maximum: 5
        includePraise:
          type: boolean
        includeActionItems:
          type: boolean
      additionalProperties: false
    CommentProfileListResponse:
      type: object
      properties:
        activeProfileId:
          type: string
          nullable: true
        items:
          type: array
          items:
            $ref: '#/components/schemas/CommentProfile'
      required: [activeProfileId, items]
...
    UsageSummary:
      type: object
      properties:
        month:
          type: string
        calls:
          type: integer
        inputTokens:
          type: integer
        outputTokens:
          type: integer
        estimatedCostUsd:
          type: number
        officialCostUsd:
          type: number
          nullable: true
        todayCalls:
          type: integer
      required:
        - month
        - calls
        - inputTokens
        - outputTokens
        - estimatedCostUsd
        - todayCalls
    BudgetConfig:
      type: object
      properties:
        monthlyBudgetUsd:
          type: number
        alertThresholds:
          type: array
          items:
            type: integer
        hardStopAtBudget:
          type: boolean
        autoDowngradeNearBudget:
          type: boolean
        autoDowngradeThresholdPct:
          type: integer
        cooldownSeconds:
          type: integer
        lastAlertThreshold:
          type: integer
          nullable: true
        usage:
          type: object
          properties:
            month:
              type: string
            calls:
              type: integer
            inputTokens:
              type: integer
            outputTokens:
              type: integer
            estimatedCostUsd:
              type: number
            spendPercent:
              type: number
            nearBudget:
              type: boolean
            overBudget:
              type: boolean
      required:
        - monthlyBudgetUsd
        - alertThresholds
        - hardStopAtBudget
        - autoDowngradeNearBudget
        - autoDowngradeThresholdPct
        - cooldownSeconds
    BudgetConfigPatchRequest:
      type: object
      properties:
        monthlyBudgetUsd:
          type: number
        alertThresholds:
          type: array
          items:
            type: integer
        hardStopAtBudget:
          type: boolean
        autoDowngradeNearBudget:
          type: boolean
        autoDowngradeThresholdPct:
          type: integer
        cooldownSeconds:
          type: integer
      additionalProperties: false
...